requirements.txt
strands-agents
strands-agents-tools
boto3
bedrock-agentcore
uvicorn[standard]
pandas
openpyxl
psycopg2-binary
numpy
fastapi